Move DirFile to proto/fs
