+++ server/http.go
@@ -16,7 +16,6 @@ import (
 	charm "github.com/charmbracelet/charm/proto"
 	"github.com/charmbracelet/charm/server/db"
 	"github.com/charmbracelet/charm/server/storage"
-	lfs "github.com/charmbracelet/charm/server/storage/local"
 	"github.com/meowgorithm/babylogger"
 	"goji.io"
 	"goji.io/pat"
@@ -268,7 +267,7 @@ func (s *HTTPServer) handleGetFile(w http.ResponseWriter, r *http.Request) {
 	}
 
 	switch f.(type) {
-	case *lfs.DirFile:
+	case *charm.DirFile:
 		w.Header().Set("Content-Type", "application/json")
 	default:
 		w.Header().Set("Content-Type", "application/octet-stream")
